Simplify and fix artifacts and builder.
The ArtifactsBuilderFor wasn't really providing a lot of reusability,
and lead to strange API docs (with artifacts). Also removed IArtifact as
it was not really needed, and changed Artifacts to make the exceptions
in the specific implementations, so that they make more sense to
developers who don't really deal with artifacts directly.